+++ likemindschat/src/main/java/com/likeminds/likemindschat/chatroom/model/GetUnreadConversationsCountResponse.kt
@@ -1,6 +1,6 @@
 package com.likeminds.likemindschat.chatroom.model
 
 data class GetUnreadConversationsCountResponse(
-    val unreadGroupChatroomConversations: Long,
-    val unreadDMChatroomConversations: Long
+    val unreadGroupChatroomConversations: Int,
+    val unreadDMChatroomConversations: Int
 )
